requirements.txt
# Third-party libraries required by the MZIC project
pyserial
pyvisa
numpy
matplotlib
Pillow
sympy
customtkinter
tkinter 
qontrol
